get rid of delete_buffer; the caller can drop a buffer at any time
so there's no need for a special method to do it. Also, this method
presumably breaks things if you accidentally delete a buffer that is
in the buffer stack but isn't the current buffer.

The detail about not deallocating the buffer contents if we don't own
it can be handled at the type level in Rust; currently we just have an
owned Vec<u8> for that, but I can go back and make it generic, with a
bound of AsRef<&[u8]>, so that we can handle both the owned and
unowned cases automatically.