WIP: adding a Rust target
This adds the Rust backend, along with a command-line switch to choose
it over the cpp backend. The rust directory contains a manual port of
a generated lexer to Rust; once it compiles and works I'll be able to
make the corresponding changes to the rust-flex.skl and rust_backend.c.